+++ plans/reviews/archive/phase-40-milestone-40-4-claude-opus-review-pass-2.md
@@ -1,0 +1,9 @@
+**Non-actionable observations**
+
+- `verification/areas/distribution_release/governance/qualification_fixture.py` is 880/900 lines (grew by 1 this round); `verification/runner/sifr_verify/selftest.py` is 860. Neither is at the cap, but both are the next refactor candidates.
+- `FORBIDDEN_CLAIMS` entries like `"-preview.2"` and `"one immutable preview version"` are exact literals rather than patterns — they catch the two pages round 1 named, but a newly written page could reintroduce preview framing in different words.
+- `planner.py:12` imports `.editor_qualification` before `.common`, out of alphabetical order with the surrounding block.
+- `plans/issues/active/phase-40-stable-channel-ga-execution.md:103-124` records no entry for round 1 or for the archived editor-pointer / vscode-package review passes, unlike the milestone_40_3 pattern; all milestone_40_4 checkboxes correctly remain unchecked, so the milestone cannot close on this diff.
+- The user-authorized deferral is recorded consistently across `plans/phases/adhoc_packaged_candidate_generated_rust.md`, `plans/phases/index.md:51`, the phase DoD, and the issue plan; the ad hoc record's evidence and hypothesis now match what the probe actually showed.
+
+Findings remain, so this round is not approved.
